The author of the Module Office Unit article emailed me to ask me to remove the link to their website
They felt that I was claiming that I wrote the article. I don't really get that, but I also don't wany anyone to be upset. He said I was a scammer, but I don't really get what the scam would be?

It does make me think a bit about my personal website. I found the link originall on https://www.swiss-miss.com/2022/12/gasp-this-modular-space.html which is a blog I've been following for a very long time.